Move storage methods to BaseStorage so it's accessed by all children and fix MRO issues.
